+++ demisto_sdk/commands/common/constants.py
@@ -1939,3 +1939,12 @@ class ImagesFolderNames(str, Enum):
 
 class PreCommitModes(str, Enum):
     NIGHTLY = "nightly"
+
+
+class InvestigationPlaybookState(str, Enum):
+    NEW = "new"  # indicates that playbook not executed yet
+    IN_PROGRESS = "inprogress"  # indicates that playbook in progress
+    PAUSED = "paused"  # indicates that playbook paused
+    COMPLETED = "completed"  # indicates that playbook completed
+    FAILED = "failed"  # indicates that playbook failed
+    WAITING = "waiting"  # indicates that playbook currently stopped and waiting for user input on manual task
